requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Advanced ML Libraries
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Configuration
PyYAML==6.0.1

# Model Serialization
joblib==1.3.2

# Progress Tracking
tqdm==4.66.1

# Utilities
python-dotenv==1.0.0